+++ tests/js/nested-list-helper.js
@@ -9,6 +9,7 @@ const realmName = process.argv[4];
 const realmModule = process.argv[5];
 
 const Realm = require(realmModule);
+
 // Ensure that schemas.js gets the correct module with `require('realm')`
 require.cache[require.resolve('realm')] = require.cache[require.resolve(realmModule)];
 let schemas = require(process.argv[2]);
@@ -44,18 +45,10 @@ function createObjects(user) {
 
     console.log("JSON: " + JSON.stringify(realm.objects('ParentObject')));
 
-    let session = realm.syncSession;
-    return new Promise((resolve, reject) => {
-        let callback = (transferred, total) => {
-            if (transferred === total) {
-                session.removeProgressNotification(callback);
-                resolve(realm);
-            }
-        }
-        session.addProgressNotification('upload', 'forCurrentlyOutstandingWork', callback);
-    });
+    return realm.syncSession.uploadAllLocalChanges();
 }
 
+// seems like we can't just use the test-utils.getRegularUser method
 const credentials = Realm.Sync.Credentials.usernamePassword(username, 'password');
 Realm.Sync.User.login('http://127.0.0.1:9080', credentials)
     .catch((error) => {